pyproject.toml
[tool.poetry]
name = "pydevts"
version = "0.1.1"
description = "PYDEVTS (PYthon Distributed EVenT System) is a distributed event system written in python and based on the concept of nodes that is designed for the implementation of data stores and other replicated systems."
authors = ["Riley Wilton <meetingprince34@outlook.com>"]
readme = "README.md"
repository = "https://github.com/Peperworx/pydevts"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
anyio = "^3.3.0"
msgpack = "^1.0.2"
pycryptodomex = "^3.10.1"


[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"